Potential fix for code scanning alert no. 21: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type size_t to avoid overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in the size_t type, which has a larger range than unsigned int.

The specific change involves casting subfaixbox->nmax to size_t before the multiplication on line 100.

+++ libopenjpeg-2.3.1/src/lib/openjpip/faixbox_manager.c
@@ -70,7 +70,7 @@ faixbox_param_t * gene_faixbox(box_param_t *box)
         subfaixbox->nmax = fetch_DBox8bytebigendian(box, (pos += 8) - 8);
         subfaixbox->m    = fetch_DBox8bytebigendian(box, (pos += 8) - 8);
 
-        numOfelem = subfaixbox->nmax * subfaixbox->m;
+        numOfelem = (size_t)subfaixbox->nmax * subfaixbox->m;
 
         subfaixbox->elem = (faixelem8_param_t *)malloc(numOfelem * sizeof(
                                faixelem8_param_t));
@@ -97,7 +97,7 @@ faixbox_param_t * gene_faixbox(box_param_t *box)
         subfaixbox->nmax = fetch_DBox4bytebigendian(box, (pos += 4) - 4);
         subfaixbox->m    = fetch_DBox4bytebigendian(box, (pos += 4) - 4);
 
-        numOfelem = subfaixbox->nmax * subfaixbox->m;
+        numOfelem = (size_t)subfaixbox->nmax * subfaixbox->m;
 
         subfaixbox->elem = (faixelem4_param_t *)malloc(numOfelem * sizeof(
                                faixelem4_param_t));
